+++ pkg/frontend/querymiddleware/querysharding_test_utils_test.go
@@ -334,7 +334,7 @@ func newEngineForTesting(t *testing.T, engine string, opts ...engineOpt) (promql
 	case querier.MimirEngine:
 		limits := streamingpromql.NewStaticQueryLimitsProvider(0)
 		metrics := stats.NewQueryMetrics(promOpts.Reg)
-		planner, err := streamingpromql.NewQueryPlanner(mqeOpts)
+		planner, err := streamingpromql.NewQueryPlanner(mqeOpts, streamingpromql.NewMaximumSupportedVersionQueryPlanVersionProvider())
 		require.NoError(t, err)
 		eng, err := streamingpromql.NewEngine(mqeOpts, limits, metrics, planner)
 		if err != nil {